django/requirements.txt
asgiref==3.2.10
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
defusedxml==0.6.0
Django==2.0.13
django-allauth==0.42.0
django-crispy-forms==1.9.2
django-gmailapi-backend==0.1.0
django-haystack==2.8.1
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.4
httplib2==0.18.1
idna==2.10
joblib==0.17.0
mod-wsgi==4.7.1
nltk==3.5
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycorenlp==0.3.0
pysolr==3.9.0
python3-openid==3.2.0
pytz==2020.1
regex==2020.9.27
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
tqdm==4.50.0
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.1.0
